fix: Improve hover functionality for task and KPI sections in ActionItem and GoalItem components
There is an issue. For items that are nested in other items (Tasks inside Action, KPIs inside Goal), if there are no Task or KPIs inside a Action or Goal, the container is not there. So input field to create new Tasks or KPIs will not be shown.

Show the input to add Tasks or KPIs when hovered over its parent Action or Goal item also.