feat: UX improvement when goal/strategy sections are empty.
In the Goals and Strategy sections, we are showing some message like it is empty when there are not goals/strategies.
This message should not be shown when the input to add new goal/strategy is visible.
In read only mode, the messages should always be shown if there is no items.